requirements.txt
torch
torchvision
scikit-learn
opencv-python
scikit-image
streamlit
numpy
Pillow
